Add files via upload
Spark - Level 1 : Issued by IBM
This badge earner has a basic understanding of Spark. The earner can describe Spark, articulate its benefits, and describe how it is used. The individual can also use Resilient Distributed Datasets (RDD) and DataFrames to perform in-memory computing and create applications on top of the Spark built-in libraries.