Add files via upload
Hadoop Foundations - Level 1
This badge earner has a basic understanding of Hadoop. The earner can describe what Big Data is and the need for Hadoop to be able to process that data in a timely manner. The individual can describe the Hadoop architecture and how to work with the Hadoop Distributed File System (HDFS) using IBM BigInsights.